test: move helper tests to FcmMessageTest as requested
