+++ tests/FcmMessageTest.php
@@ -109,4 +109,43 @@ class FcmMessageTest extends TestCase
 
         $this->assertSame($client, $message->client);
     }
+
+    public function test_appends_android_options_into_custom()
+    {
+        $msg = FcmMessage::create()
+            ->notification(new Notification(title: 'T', body: 'B'))
+            ->android(['notification' => ['channel_id' => 'ops', 'sound' => 'default']]);
+
+        $payload = $msg->toArray();
+
+        $this->assertArrayHasKey('android', $payload);
+        $this->assertArrayHasKey('notification', $payload['android']);
+        $this->assertEquals('ops', $payload['android']['notification']['channel_id']);
+        $this->assertEquals('default', $payload['android']['notification']['sound']);
+    }
+
+    public function test_appends_ios_options_into_custom()
+    {
+        $msg = FcmMessage::create()
+            ->ios(['payload' => ['aps' => ['sound' => 'default']]]);
+
+        $payload = $msg->toArray();
+
+        $this->assertArrayHasKey('apns', $payload);
+        $this->assertArrayHasKey('payload', $payload['apns']);
+        $this->assertEquals('default', $payload['apns']['payload']['aps']['sound']);
+    }
+
+    public function test_preserves_existing_custom_keys_when_using_helpers()
+    {
+        $msg = FcmMessage::create()
+            ->custom(['meta' => ['a' => 1]])
+            ->android(['notification' => ['color' => '#000']]);
+
+        $payload = $msg->toArray();
+
+        $this->assertArrayHasKey('meta', $payload);
+        $this->assertEquals(1, $payload['meta']['a']);
+        $this->assertEquals('#000', $payload['android']['notification']['color']);
+    }
 }
